Add files via upload
This project is based on Dry bean Dataset, which is extracted from Kaggle. This project is for obtaining the uniform seed classification out of 7 different dry bean classes. The target variable is 'Class' column. I have used different libraries related to python and Machine Learning- supervised learning models for prediction and to find the accuracy related to actual VS predicted values.